eval network

+++ chess4j-java/src/main/java/com/jamesswafford/chess4j/eval/Evaluator.java
@@ -6,6 +6,6 @@ public interface Evaluator {
 
     int evaluateBoard(Board board);
 
-    double evaluateBoardWithNN(Board board);
+    int evaluateBoardWithNN(Board board);
 
 }
